merge

+++ setup.sh
@@ -1,12 +1,27 @@
 #!/usr/bin/env bash
 #encoding=utf8
 
+function check_installed_pip() {
+   ${PYTHON} -m pip > /dev/null
+   if [ $? -ne 0 ]; then
+        echo "pip not found (called as '${PYTHON} -m pip'). Please make sure that pip is available for ${PYTHON}."
+        exit 1
+   fi
+}
+
 # Check which python version is installed
 function check_installed_python() {
+    if [ -n "${VIRTUAL_ENV}" ]; then
+        echo "Please deactivate your virtual environment before running setup.sh."
+        echo "You can do this by running 'deactivate'."
+        exit 2
+    fi
+
     which python3.7
     if [ $? -eq 0 ]; then
         echo "using Python 3.7"
         PYTHON=python3.7
+        check_installed_pip
         return
     fi
 
@@ -14,6 +29,7 @@ function check_installed_python() {
     if [ $? -eq 0 ]; then
         echo "using Python 3.6"
         PYTHON=python3.6
+        check_installed_pip
         return
    fi
 
@@ -21,29 +37,30 @@ function check_installed_python() {
         echo "No usable python found. Please make sure to have python3.6 or python3.7 installed"
         exit 1
    fi
-
 }
 
 function updateenv() {
     echo "-------------------------"
     echo "Updating your virtual env"
     echo "-------------------------"
+    if [ ! -f .env/bin/activate ]; then
+        echo "Something went wrong, no virtual environment found."
+        exit 1
+    fi
     source .env/bin/activate
-    echo "pip3 install in-progress. Please wait..."
-    # Install numpy first to have py_find_1st install clean
-    pip3 install --upgrade pip numpy
-    pip3 install --upgrade -r requirements.txt
-
+    echo "pip install in-progress. Please wait..."
+    ${PYTHON} -m pip install --upgrade pip
     read -p "Do you want to install dependencies for dev [y/N]? "
     if [[ $REPLY =~ ^[Yy]$ ]]
     then
-        pip3 install --upgrade -r requirements-dev.txt
+        ${PYTHON} -m pip install --upgrade -r requirements-dev.txt
     else
+        ${PYTHON} -m pip install --upgrade -r requirements.txt
         echo "Dev dependencies ignored."
     fi
 
-    pip3 install --quiet -e .
-    echo "pip3 install completed"
+    ${PYTHON} -m pip install -e .
+    echo "pip install completed"
     echo
 }
 
@@ -61,6 +78,10 @@ function install_talib() {
     ./configure --prefix=/usr/local
     make
     sudo make install
+    if [ -x "$(command -v apt-get)" ]; then
+        echo "Updating library path using ldconfig"
+        sudo ldconfig
+    fi
     cd .. && rm -rf ./ta-lib/
     cd ..
 }
@@ -74,16 +95,14 @@ function install_macos() {
         echo "-------------------------"
         /usr/bin/ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
     fi
-    brew install python3 wget
     install_talib
     test_and_fix_python_on_mac
 }
 
 # Install bot Debian_ubuntu
 function install_debian() {
-    sudo add-apt-repository ppa:jonathonf/python-3.6
     sudo apt-get update
-    sudo apt-get install python3.6 python3.6-venv python3.6-dev build-essential autoconf libtool pkg-config make wget git
+    sudo apt-get install -y build-essential autoconf libtool pkg-config make wget git
     install_talib
 }
 
@@ -98,30 +117,39 @@ function reset() {
     echo "----------------------------"
     echo "Reseting branch and virtual env"
     echo "----------------------------"
+
     if [ "1" == $(git branch -vv |grep -cE "\* develop|\* master") ]
     then
-        if [ -d ".env" ]; then
-          echo "- Delete your previous virtual env"
-          rm -rf .env
-        fi
 
-        git fetch -a
+        read -p "Reset git branch? (This will remove all changes you made!) [y/N]? "
+        if [[ $REPLY =~ ^[Yy]$ ]]; then
 
-        if [ "1" == $(git branch -vv |grep -c "* develop") ]
-        then
-          echo "- Hard resetting of 'develop' branch."
-          git reset --hard origin/develop
-        elif [ "1" == $(git branch -vv |grep -c "* master") ]
-        then
-          echo "- Hard resetting of 'master' branch."
-          git reset --hard origin/master
+            git fetch -a
+
+            if [ "1" == $(git branch -vv |grep -c "* develop") ]
+            then
+                echo "- Hard resetting of 'develop' branch."
+                git reset --hard origin/develop
+            elif [ "1" == $(git branch -vv |grep -c "* master") ]
+            then
+                echo "- Hard resetting of 'master' branch."
+                git reset --hard origin/master
+            fi
         fi
     else
         echo "Reset ignored because you are not on 'master' or 'develop'."
     fi
 
+    if [ -d ".env" ]; then
+        echo "- Delete your previous virtual env"
+        rm -rf .env
+    fi
     echo
     ${PYTHON} -m venv .env
+    if [ $? -ne 0 ]; then
+        echo "Could not create virtual environment. Leaving now"
+        exit 1
+    fi
     updateenv
 }
 
@@ -235,7 +263,7 @@ function install() {
     echo "-------------------------"
     echo "Run the bot !"
     echo "-------------------------"
-    echo "You can now use the bot by executing 'source .env/bin/activate; python freqtrade/main.py'."
+    echo "You can now use the bot by executing 'source .env/bin/activate; freqtrade'."
 }
 
 function plot() {
@@ -244,7 +272,7 @@ echo "
 Installing dependencies for Plotting scripts
 -----------------------------------------
 "
-pip install plotly --upgrade
+${PYTHON} -m pip install plotly --upgrade
 }
 
 function help() {
